+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/MiseAjourDeSite.labels.js
