+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/MiseAjourDeSite.schema.js
@@ -1,5 +1,5 @@
 import { object, string, number, array, ref } from "yup";
-import labels from "./MiseAjourSectionDeSite.labels";
+import labels from "./MiseAjourDeSite.labels";
 
 function emptyStringToNull(value, originalValue) {
     if (typeof originalValue === "string" && originalValue === "") {
