+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/inputs/InputElectricityAccess.vue
