+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/inputs/InputPestAnimalsComments.vue
@@ -12,5 +12,5 @@
 
 <script setup>
 import { TextArea } from "@resorptionbidonvilles/ui";
-import labels from "#src/js/pages/MiseAjourSectionDeSite/MiseAJourSectionDeSite.labels";
+import labels from "../MiseAJourDeSite.labels";
 </script>
