+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/inputs/InputSanitaryToiletsAreInside.vue
@@ -1,8 +1,8 @@
 <template>
-    <SubQuestionWrapper :label="labels.sanitary_open_air_defecation">
+    <SubQuestionWrapper :label="labels.sanitary_toilets_are_inside">
         <CheckableGroup
-            id="sanitary_open_air_defecation"
-            direction="vertical"
+            id="sanitary_toilets_are_inside"
+            direction="horizontal"
             withoutMargin
         >
             <Radio
@@ -10,7 +10,7 @@
                 :key="item.value"
                 :value="item.value"
                 :label="item.label"
-                name="sanitary_open_air_defecation"
+                name="sanitary_toilets_are_inside"
                 v-bind="$attrs"
                 variant="radio"
                 allowNull
@@ -27,7 +27,7 @@ import {
     Radio,
     SubQuestionWrapper,
 } from "@resorptionbidonvilles/ui";
-import labels from "#src/js/pages/MiseAjourSectionDeSite/MiseAJourSectionDeSite.labels";
+import labels from "../MiseAJourDeSite.labels";
 
 const items = [
     { value: 1, label: "Oui" },
