+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/inputs/InputSanitaryWorkingToilets.vue
