+++ packages/frontend/mobile/src/js/components/FormMiseAJourSectionDeSite/inputs/InputWaterAccessType.vue
